backend/requirements.txt
fastapi
uvicorn
openai
requests
python-dotenv
pyodbc
pandas
yfinance
pymupdf
chromadb
tiktoken
langchain
